Move `cloneFile` and `renameFile` over
