+++ packages/interface/README.md
@@ -30,6 +30,7 @@ _Returns_
 
 -   `OpenModalAction`: Action object.
 
+
 <!-- END TOKEN(Autogenerated actions|src/store/actions.ts) -->
 
 ### Selectors
@@ -50,6 +51,5 @@ _Parameters_
 _Returns_
 
 -   `boolean`: Whether the modal is active.
-
 
 <!-- END TOKEN(Autogenerated selectors|src/store/selectors.ts) -->
