+++ packages/media-recording/README.md
@@ -109,7 +109,6 @@ Returns an action object signalling that audio mode should be toggled.
 
 Updates the list of available media devices.
 
-
 <!-- END TOKEN(Autogenerated actions|src/store/actions.ts) -->
 
 ### Selectors
@@ -296,5 +295,6 @@ Whether the app is currently in recording mode.
 _Parameters_
 
 -   _state_ `State`: Recording state.
+
 
 <!-- END TOKEN(Autogenerated selectors|src/store/selectors.ts) -->
