+++ packages/media-utils/README.md
@@ -6,18 +6,6 @@ A set of media-related helper utilities.
 
 <!-- START TOKEN(Autogenerated API docs) -->
 
-### cloneFile
-
-Clones a given file object.
-
-_Parameters_
-
--   _file_ `File`: File object.
-
-_Returns_
-
--   `File`: New file object.
-
 ### getFileBasename
 
 Returns file basename without extension.
@@ -53,21 +41,6 @@ Small wrapper around the `File` class to hold information about current dimensio
 _Type_
 
 -   `ImageFile`
-
-### renameFile
-
-Renames a given file and returns a new file.
-
-Copies over the last modified time.
-
-_Parameters_
-
--   _file_ `File`: File object.
--   _name_ `string`: File name.
-
-_Returns_
-
--   `File`: Renamed file object.
 
 
 <!-- END TOKEN(Autogenerated API docs) -->
