+++ packages/media-utils/src/index.ts
@@ -23,29 +23,3 @@ export function getFileBasename( name: string ): string {
 export function getFileExtension( file: string ): string | null {
 	return file.includes( '.' ) ? file.split( '.' ).pop() || null : null;
 }
-
-/**
- * Renames a given file and returns a new file.
- *
- * Copies over the last modified time.
- *
- * @param file File object.
- * @param name File name.
- * @return Renamed file object.
- */
-export function renameFile( file: File, name: string ): File {
-	return new File( [ file ], name, {
-		type: file.type,
-		lastModified: file.lastModified,
-	} );
-}
-
-/**
- * Clones a given file object.
- *
- * @param file File object.
- * @return New file object.
- */
-export function cloneFile( file: File ): File {
-	return renameFile( file, file.name );
-}
